CIRC-1072 Refactoring according to code review
(cherry picked from commit f2e07b5abaf7679c10243a8a866a151345cccadc)
