+++ src/test/java/api/support/fixtures/UserManualBlocksFixture.java
@@ -1,6 +1,9 @@
 package api.support.fixtures;
 
+import static org.folio.circulation.support.ClockManager.getClockManager;
 import static org.folio.circulation.support.json.JsonPropertyFetcher.getProperty;
+
+import java.util.UUID;
 
 import api.support.builders.UserManualBlockBuilder;
 import api.support.http.ResourceClient;
@@ -23,5 +26,21 @@ public class UserManualBlocksFixture {
   public void cleanUp() {
 
     userManualBlocksRecordCreator.cleanUp();
+  }
+
+  public void createManualPatronBlockForUser(UUID requesterId) {
+    create(getManualBlockBuilder()
+      .withRequests(true)
+      .withExpirationDate(getClockManager().getDateTime().plusYears(1))
+      .withUserId(requesterId.toString()));
+  }
+
+  private UserManualBlockBuilder getManualBlockBuilder() {
+    return new UserManualBlockBuilder()
+      .withType("Manual")
+      .withDesc("Display description")
+      .withStaffInformation("Staff information")
+      .withPatronMessage("Patron message")
+      .withId(UUID.randomUUID());
   }
 }
